Print an intX_t by casting to int64_t

+++ codec/decoder/core/src/decode_slice.cpp
@@ -415,8 +415,8 @@ int32_t WelsDecodeSlice (PWelsDecoderContext pCtx, bool bFirstSliceInLayer, PNal
     }
     if (iUsedBits > pBs->iBits) { //When BS incomplete, as long as find it, SHOULD stop decoding to avoid mosaic or crash.
       WelsLog (pCtx, WELS_LOG_WARNING,
-               "WelsDecodeSlice()::::pBs incomplete, iUsedBits:%d > pBs->iBits:%d, MUST stop decoding.\n",
-               iUsedBits, pBs->iBits);
+               "WelsDecodeSlice()::::pBs incomplete, iUsedBits:%"PRId64" > pBs->iBits:%d, MUST stop decoding.\n",
+               (int64_t) iUsedBits, pBs->iBits);
       return -1;
     }
     iMbX = iNextMbXyIndex % pCurLayer->iMbWidth;
